requirements_optimized.txt
# Core dependencies (existing)
openai>=1.0.0
pinecone-client>=3.0.0
anthropic>=0.8.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pandas>=2.0.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
streamlit>=1.28.0

# Optimization dependencies (new)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
redis>=5.0.0
hiredis>=2.2.0

# Optional for production deployment
gunicorn>=21.2.0  # Production WSGI server
python-multipart>=0.0.6  # For form data
pydantic>=2.0.0  # Data validation
